move samples from commonjs to mjs
