+++ samples/src/node/gen/waitforfullsize.mjs
@@ -8,12 +8,12 @@ it. If you have received this file from a source other than Adobe,
 then your use, modification, or distribution of it requires the prior
 written permission of Adobe. 
 */
-const LrSession = require('../../common/lr/LrSession')
+import LrSession from '../../common/lr/LrSession.mjs'
 
 async function mainP(assetId) {
 	let lr = await LrSession.currentContextP()
-	let response = await lr.generate2560RenditionP(assetId)
-	console.log(JSON.stringify(response, null, 2))
+	let result = await lr.waitForFullsizeRendtionP(assetId)
+	console.log('result:', result)
 }
 
 mainP(process.argv[2]).then(() => console.log('done')).catch(e => console.error('error:', e))
