+++ samples/src/node/get/getfirstalbumasset.mjs
@@ -8,13 +8,12 @@ it. If you have received this file from a source other than Adobe,
 then your use, modification, or distribution of it requires the prior
 written permission of Adobe. 
 */
-const LrSession = require('../../common/lr/LrSession')
-const LrUtils = require('../../common/lr/LrUtils')
+import LrSession from '../../common/lr/LrSession.mjs'
 
-async function mainP() {
+async function mainP(albumId) {
 	let lr = await LrSession.currentContextP()
-	let root = await LrUtils.getRootProjectSetP(lr)
-	await LrUtils.logAlbumHierarchyP(lr, root)
+	let response = await lr.getFirstAlbumAssetP(albumId)
+	console.log(JSON.stringify(response, null, 2))
 }
 
-mainP().then(() => console.log('done')).catch(e => console.error('error:', e))
+mainP(process.argv[2]).then(() => console.log('done')).catch(e => console.error('error:', e))
