+++ samples/src/node/get/getincompletes.mjs
@@ -8,12 +8,12 @@ it. If you have received this file from a source other than Adobe,
 then your use, modification, or distribution of it requires the prior
 written permission of Adobe. 
 */
-const LrSession = require('../../common/lr/LrSession')
+import LrSession from '../../common/lr/LrSession.mjs'
 
-async function mainP(albumId) {
+async function mainP() {
 	let lr = await LrSession.currentContextP()
-	let response = await lr.getFirstAlbumAssetP(albumId)
+	let response = await lr.getIncompletesP()
 	console.log(JSON.stringify(response, null, 2))
 }
 
-mainP(process.argv[2]).then(() => console.log('done')).catch(e => console.error('error:', e))
+mainP().then(() => console.log('done')).catch(e => console.error('error:', e))
