+++ samples/src/node/upload/uploadfile.mjs
@@ -8,9 +8,9 @@ it. If you have received this file from a source other than Adobe,
 then your use, modification, or distribution of it requires the prior
 written permission of Adobe. 
 */
-const LrSession = require('../../common/lr/LrSession')
-const File = require('../../common/file/File')
-const FileUtils = require('../../common/file/FileUtils')
+import LrSession from '../../common/lr/LrSession.mjs'
+import File from '../../common/file/File.mjs'
+import FileUtils from '../../common/file/FileUtils.mjs'
 
 async function mainP(filePath) {
 	let lr = await LrSession.currentContextP()
